20 - Add different checks from the user role

+++ src/controllers/post/DeletePostController.php
@@ -8,9 +8,9 @@ use App\lib\DatabaseConnection;
 use App\lib\PostIdChecker;
 use App\lib\SessionChecker;
 use App\lib\SessionManager;
+use App\Lib\UserChecker;
 use App\lib\View;
 use App\model\PostRepository;
-use Exception;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
@@ -32,13 +32,25 @@ class DeletePostController
 
         $sessionChecker->sessionChecker();
         $sessionData = $sessionChecker->getSessionData();
+        $userChecker = new UserChecker();
+        $error = null;
 
         $view = new View();
         $postId = PostIdChecker::getId($args);
         $post = $this->getPostsRepository()->getPost($postId);
 
-        $html = $view->render('post_delete.twig', ['post' => $post, 'session' => $sessionData]);
-        $response->getBody()->write($html);
+        if (
+            ($userChecker->isAuthenticated($sessionData['token'] ?? '') === true
+            && $userChecker->isCurrentUser($post->userId, $sessionData['id']) === true)
+            || $userChecker->isAdmin($sessionData['role'] ?? 'ROLE_USER')
+        ) {
+            $html = $view->render('post_delete.twig', ['post' => $post, 'session' => $sessionData, 'error' => $error]);
+            $response->getBody()->write($html);
+        } else {
+            $error = 'Vous n\'avez pas accès à cette page !';
+
+            return $this->renderErrorResponse($response, $error);
+        }
 
         return $response;
     }
@@ -54,28 +66,64 @@ class DeletePostController
      */
     public function remove(RequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
-        if ($request->getMethod() === 'POST') {
-            $postRepository = $this->getPostsRepository();
-            $postId = PostIdChecker::getId($args);
-            $error = null;
+        $sessionManager = new SessionManager();
+        $sessionChecker = new SessionChecker($sessionManager);
 
-            $success = $postRepository->deletePost($postId);
+        $sessionChecker->sessionChecker();
+        $sessionData = $sessionChecker->getSessionData();
+        $userChecker = new UserChecker();
 
-            if ($success === false) {
-                $error = 'Une erreur est survenue dans la suppression de l\'article.';
-            } else {
-                return $response->withHeader('Location', '/blog')->withStatus(302);
+        $postRepository = $this->getPostsRepository();
+        $postId = PostIdChecker::getId($args);
+        $fetchPost = $postRepository->getPost($postId);
+
+        $error = null;
+        $view = new View();
+        if (
+            ($userChecker->isAuthenticated($sessionData['token'] ?? '')
+                && $userChecker->isCurrentUser($fetchPost->userId, $sessionData['id']))
+            || $userChecker->isAdmin($sessionData['role'] ??  'ROLE_USER')
+        ) {
+            if ($request->getMethod() === 'POST') {
+                $success = $postRepository->deletePost($postId);
+
+                if ($success === false) {
+                    $error = 'Une erreur est survenue dans la suppression de l\'article.';
+                } else {
+                    return $response->withHeader('Location', '/blog')->withStatus(302);
+                }
+
+                $view = new View();
+                $html = $view->render('post_delete.twig', ['error' => $error]);
+
+                $response->getBody()->write($html);
+
+                return $response;
             }
+        } else {
+            $error = 'Vous n\'avez pas accès à cette page !';
 
-            $view = new View();
-            $html = $view->render('post_delete.twig', ['error' => $error]);
-
-            $response->getBody()->write($html);
-
-            return $response;
+            return $this->renderErrorResponse($response, $error);
         }
 
-        throw new Exception('Une erreur est survenue');
+        return $response;
+    }
+
+    /**
+     * renderErrorResponse
+     *
+     * @param  ResponseInterface $response
+     * @param  string $error
+     *
+     * @return ResponseInterface
+     */
+    private function renderErrorResponse(ResponseInterface $response, string $error): ResponseInterface
+    {
+        $view = new View();
+        $html = $view->render('error.twig', ['error' => $error]);
+        $response->getBody()->write($html);
+
+        return $response;
     }
 
     /**
